Get element from UI state and not from service
We are using positions here. We could either give up on positions and search from the string value
in the widget for the actual one in the Service, or give the UI its own vector of Posts, on
which it can rightfully access positions. This commit takes the second approa This commit takes the second approach.